2.1.0

+++ SoraPassport/Common/Model/Chain.swift
@@ -1,12 +1,7 @@
-/**
-* Copyright Soramitsu Co., Ltd. All Rights Reserved.
-* SPDX-License-Identifier: Apache 2.0
-*/
-
 import Foundation
 import IrohaCrypto
 
-typealias SNAddressType = UInt8
+typealias SNAddressType = UInt16
 
 enum Chain: String, Codable, CaseIterable {
     case polkadot = "Polkadot"
